Apply prs feedback
